% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStageDistribution.R
\name{calculateStageDistribution}
\alias{calculateStageDistribution}
\title{Calculate Observed Stage Distribution}
\usage{
calculateStageDistribution(dataset, n.seeds)
}
\arguments{
\item{dataset}{}
}
\description{
Calculate Observed Stage Distribution
}

