% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTransitionRates.R
\name{calculateTransitionRates}
\alias{calculateTransitionRates}
\title{Create Transition Rate Matrix}
\usage{
calculateTransitionRates(transition_counts, n_per_stage)
}
\arguments{
\item{transition_counts}{Matrix of transition counts (representing the number of individuals observed to transition from the stage represented by the column name to the stage represented by the row name).}

\item{n_per_stage}{Observed number of individuals per stage}
}
\value{
Returns A Matrix of the same dimensions as transition_counts that gives the rates at which individuals transition from one stage to another. This is calculated as the number in a given transition_counts cell divided by the number of individuals observed in that stage (the stage given by the column name of that transition_counts cell).
}
\description{
Create Transition Rate Matrix
}

