% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifybyStage.R
\name{classifybyStage}
\alias{classifybyStage}
\title{Classify plants by Stage}
\usage{
classifybyStage(Data)
}
\arguments{
\item{Data}{}
}
\description{
Classify plants by stage. Individuals remain a seedling only for one year, then automatically transition to a "juvenile" stage. Once a juvenile plant reaches two segments in size, it is no longer a juvenile. Once a plant reaches the "Adult" stage, they remain "Adult", even if they retrogress to one segment.
}

