% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClonalReproTransitionMatrix.R
\name{createClonalReproTransitionMatrix}
\alias{createClonalReproTransitionMatrix}
\title{Create Clonal Reproduction Transition Matrix}
\usage{
createClonalReproTransitionMatrix(clonal_repro_dataset, TMdata)
}
\arguments{
\item{clonal_repro_dataset}{This dataset is created by the calculateClonalReproduction function. Each row represents information for a unique offspring. Each row contains the identity of the parent, the year, the size of the offspring, and the size of the parent.}

\item{TMdata}{Survey data used to create transition matrix.}
}
\value{
Returns a list in which the first element is the count of the number of individuals in each stage-stage transition and the second element contains the rates at which individuals transition among stages.
}
\description{
Create Clonal Reproduction Transition Matrix
}

