% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDeadFactorVariable.R
\name{createDeadFactorVariable}
\alias{createDeadFactorVariable}
\title{Create Dead Variable (Factor Type)}
\usage{
createDeadFactorVariable(Data)
}
\arguments{
\item{Data}{Dataset}
}
\description{
Create Dead Variable (Factor Type)
}

