% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHabitatType.R
\name{createHabitatType}
\alias{createHabitatType}
\title{Process Create Habitat Type Variable}
\usage{
createHabitatType(dat)
}
\arguments{
\item{data}{}
}
\description{
Process Create Habitat Type Variable
}

