% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInsectPresDuringStudy.R
\name{createInsectPresDuringStudy}
\alias{createInsectPresDuringStudy}
\title{Make Insect Factor Variables}
\usage{
createInsectPresDuringStudy(Plant_Surveys_by_Year)
}
\arguments{
\item{Plant_Surveys_by_Year}{Plant Survey Dataset}
}
\description{
Make new insect variables with "yes"/"no" values instead of 0/1.
}

