% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInsectCategoricalVariable.R
\name{createMothOccupancytVariable}
\alias{createMothOccupancytVariable}
\title{Create Moth Occupancy Variable}
\usage{
createMothOccupancytVariable()
}
\arguments{
\item{Plant_Surveys_by_Year}{Dataset}
}
\description{
Create Moth Occupancy Variable
}

