% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInsectCategoricalVariable.R
\name{createMothOccupancyVariable}
\alias{createMothOccupancyVariable}
\title{Create Moth Occupancy During Study Variable}
\usage{
createMothOccupancyVariable(Plant_Surveys_by_Year)
}
\arguments{
\item{Plant_Surveys_by_Year}{Dataset}
}
\description{
Create Moth Occupancy During Study Variable
}

