% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlantInfobyPlant.R
\name{createPlantInfobyPlant}
\alias{createPlantInfobyPlant}
\title{Create Plant Info by Plant}
\usage{
createPlantInfobyPlant(Plant.Info, Plant.Surveys.by.Year,
  Plant.Surveys.by.Plant)
}
\arguments{
\item{Plant_Surveys_by_Plant}{Plant Survey Dataset}
}
\description{
Create variables giving evidence of moths and insects, including moth presence or damage.
}

