% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlantInsectPresence.R
\name{createPlantInsectPresence}
\alias{createPlantInsectPresence}
\title{Create Moth Evidence and Insect Evidence Variables}
\usage{
createPlantInsectPresence(Plant_Surveys_by_Plant)
}
\arguments{
\item{Plant_Surveys_by_Plant}{Plant Survey Dataset}
}
\description{
Create variables giving evidence of moths and insects, including moth presence or damage.
}

