% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlantSurveysbyYear.R
\name{createPlantSurveysbyYear}
\alias{createPlantSurveysbyYear}
\title{Create Dataset by Year}
\usage{
createPlantSurveysbyYear(Plant_Surveys_by_Plant)
}
\arguments{
\item{Plant_Surveys_by_Plant}{Plant Survey Dataset}
}
\description{
Create variables indicating whether an insect was present during the study. Useful in particular for Life Table Response Experiments.
}

