% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProjectionMatrix.R
\name{createProjectionMatrix}
\alias{createProjectionMatrix}
\title{Create Projection Matrix}
\usage{
createProjectionMatrix(trans_data, SeedBankSize, SeedsPerFruit, SeedSurvival)
}
\arguments{
\item{trans_data}{transition data}

\item{SeedBankSize}{seed bank size estimate}

\item{SeedsPerFruit}{average number of seeds per fruit}

\item{SeedSurvival}{seed survival estimate}
}
\value{
Return a projection matrix
}
\description{
Create projection matrix with growth, stasis, retrogression, survival, and sexual reproduction rates. Assumptions about the recruitment rate, seed bank, and fertilities follow Stubben (2007).
}

