% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSequenceofDates.R
\name{createSequenceofDates}
\alias{createSequenceofDates}
\title{Create sequence of dates, by fecundity year, to use as date intervals in mergePlantRecordsfromMultiplePlots}
\usage{
createSequenceofDates(seasons = seasons, probs = seq(0, 1, 0.1125))
}
\arguments{
\item{probs}{the quantiles to use to divide up the fecundity year (e.g., probs=seq(0,1,0.1125))}
}
\description{
Create sequence of dates, by fecundity year, to use as date intervals in mergePlantRecordsfromMultiplePlots
}

