% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFirstDeadMissingObservation.R
\name{findFirstDeadMissingObservation}
\alias{findFirstDeadMissingObservation}
\title{Filter Plant Surveys by Plant and Keep Only the First Dead/Missing Observation}
\usage{
findFirstDeadMissingObservation(Plant_Surveys_by_Plant)
}
\description{
Important for determining the first date the plant was marked dead or missing. Also, subsequent surveys recording death/missingness are not necessary for analysis.
}

