% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPopulationMatrices.R
\name{firstup}
\alias{firstup}
\title{from https://stackoverflow.com/questions/18509527/first-letter-to-upper-case}
\usage{
firstup(x)
}
\description{
from https://stackoverflow.com/questions/18509527/first-letter-to-upper-case
}

