% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixDeadRecords.R
\name{fixDeadRecords}
\alias{fixDeadRecords}
\title{Fix Dead Records for Transition Matrix Data}
\usage{
fixDeadRecords(Plant.Surveys.TM.largest.parent)
}
\arguments{
\item{Plant.Surveys.TM.largest.parent}{Transition matrix data.}
}
\description{
Fix Dead Records for Transition Matrix Data
}

