% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixInconsistentDeadMissingRecords.R
\name{fixInconsistentDeadMissingRecords}
\alias{fixInconsistentDeadMissingRecords}
\title{Fix Surveys for Plants with Inconsistent Records of Dead/Missing}
\usage{
fixInconsistentDeadMissingRecords(Plant_Surveys)
}
\arguments{
\item{Plant_Surveys}{Dataset with Plant Surveys}
}
\description{
Process Plant_Surveys
}

