% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotElasticitySensitivity.R
\name{plotElasticitySensitivity}
\alias{plotElasticitySensitivity}
\title{Plot Elasticity and Sensitivity Data}
\usage{
plotElasticitySensitivity(dat, x_variable, y_variable, grouping_variable,
  facet_grid_var, xlab_label, ylab_label, n_transitions_per_stage)
}
\arguments{
\item{dat}{Dataset}

\item{n_transitions_per_stage}{}
}
\description{
Plot Elasticity and Sensitivity Data
}

