% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeedBank.R
\name{plotSeedBank}
\alias{plotSeedBank}
\title{Plot Seed Bank Data}
\usage{
plotSeedBank(dat, x_variable, grouping_variable, facet_grid_var,
  legend_title = "Parent Assignment\\nMethod", xlab_label)
}
\arguments{
\item{dat}{Dataset}

\item{grouping_variable}{}

\item{legend_title}{Default value is "Parent Assignment\nMethod."}

\item{xlab_label}{}
}
\description{
Plot Seed Bank Data
}

