% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeedBankSeedSurvival.R
\name{plotSeedBankSeedSurvival}
\alias{plotSeedBankSeedSurvival}
\title{Plot Seed Bank Data}
\usage{
plotSeedBankSeedSurvival(dat, x_variable, grouping_variable, facet_grid_var,
  legend_title = "Parent Assignment\\nMethod", xlab_label, strip.text.size,
  x_breaks, x_limits, x_coord_trans, y_breaks, y_limits, strip.text.y.angle,
  axis.text.x.size, legend.text.size, legend.title.size)
}
\arguments{
\item{dat}{Dataset}

\item{grouping_variable}{}

\item{legend_title}{Default value is "Parent Assignment Method."}

\item{strip.text.size}{}
}
\description{
Plot Seed Bank Data
}

