% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepDataTransitionMatrix.R
\name{prepDataTransitionMatrix}
\alias{prepDataTransitionMatrix}
\title{Prepare Data For Transition Matrix Function}
\usage{
prepDataTransitionMatrix(Dat, clonal_repro_dataset, SizeClass, TransitionYear)
}
\arguments{
\item{Dat}{Dataset}

\item{SizeClass}{Classes into which to divide individuals based on size}

\item{TransitionYear}{Year}
}
\description{
Prepare Data For Transition Matrix Function
}

