% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPopulationMatrices.R
\name{printLTREPopulationMatrices}
\alias{printLTREPopulationMatrices}
\title{Successively print LTRE population matrices for inclusion in an appendix}
\usage{
printLTREPopulationMatrices(data.list, parent.method, years, first.year,
  LTRE_variable, LTRE_variable_section_heading)
}
\arguments{
\item{years}{}

\item{LTRE_variable}{}
}
\description{
Successively print LTRE population matrices for inclusion in an appendix
}

