% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPopulationMatrices.R
\name{printPopulationMatrices}
\alias{printPopulationMatrices}
\title{Successively print population matrices for inclusion in an appendix}
\usage{
printPopulationMatrices(data.list, parent.method, years, first.year,
  second.year)
}
\arguments{
\item{years}{}

\item{first.year}{}
}
\description{
Successively print population matrices for inclusion in an appendix
}

