% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processOccupancyPlotInfo.R
\name{processOccupancyPlotInfo}
\alias{processOccupancyPlotInfo}
\title{Process Plot Survey Data}
\usage{
processOccupancyPlotInfo(Plot_Info, Plot.Vegetation, GIS_data_updated)
}
\description{
Steps:
\itemize{
 \item fix Cluster variable name
 \item include "In Demography Study" information
 \item add vegetation information
\itemize{
 \item VegDensity
 \item Canopy
 \item Detritus
}
 \item add elevation, distance to water information
\itemize{
 \item elevation
 \item dist_to_water
 }
 \item limit to only plots still in study
 \item make coordinates numeric
	}
}

