% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processOccupancyPlotSurveys.R
\name{processOccupancyPlotSurveys}
\alias{processOccupancyPlotSurveys}
\title{Process Plot Survey Data}
\usage{
processOccupancyPlotSurveys(Plot_Surveys, Plot_Info)
}
\description{
Steps:
\enumerate{
 \item Warnings and Errors:
\itemize{
 \item Duplicate plot surveys for plot(s)
}	
 \item limit to only plots still in study
 \item format date
 \item Change Yes/No to 1/0
 \item make variables numeric
 \item If \emph{O. pusilla} is not present, change all \emph{O. pusilla}-dependent data to NA
 \item If O. stricta is not present, change all O. stricta-dependent data to NA
}
}

