% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPlantInfo.R
\name{processPlantInfo}
\alias{processPlantInfo}
\title{Process Plant Info}
\usage{
processPlantInfo(Plant.Info, Plot.Info, Plant.Surveys)
}
\arguments{
\item{Plant.Info}{Dataset with Plant Information}

\item{Plot.Info}{Dataset with Plot Information}
}
\description{
Process Plant Info
}

