% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPopulationModelResults.R
\name{processPopulationModelingResults}
\alias{processPopulationModelingResults}
\title{Process Population Model Results}
\usage{
processPopulationModelingResults(MPMresults, LTRE_variable = NULL)
}
\arguments{
\item{LTRE_variable}{Life table response variable, Default value is NULL.}
}
\description{
Prepare population modeling results for plotting.
}

