% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPopulationModels.R
\name{runPopulationModels}
\alias{runPopulationModels}
\title{Run Population Models}
\usage{
runPopulationModels(ldf, cldf, SizeClass, TransitionYear, SeedSurvival,
  SeedBankSize, SeedsPerFruit, n.iter)
}
\arguments{
\item{ldf}{list of data frames containing transition information for plants}

\item{cldf}{list of data frames containing information on clones and their parents}

\item{n.iter}{}
}
\description{
Prepare transition matrices, then project matrices and analyze results.
}

