% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceofDates.R
\name{SequenceofDates}
\alias{SequenceofDates}
\title{Datasets of sequence of dates, by fecundity year, to use as date intervals in mergePlantRecordsfromMultiplePlots}
\format{A vector of dates ranging from 2008-03-20 to 2016-02-12}
\description{
Datasets of sequence of dates, by fecundity year, to use as date intervals in mergePlantRecordsfromMultiplePlots
}

