% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateClonalReproduction.R
\name{calculateClonalReproduction}
\alias{calculateClonalReproduction}
\title{Calculate Clonal Reproduction}
\usage{
calculateClonalReproduction(Plant.Surveys.by.Year, Plant.Surveys.by.Plant,
  Plant.Info.Analysis, ParentChoice = "None")
}
\arguments{
\item{Parent.Choice}{}
}
\value{
Returns the following variables:
\itemize{
 	\item first size of offspring
 	\item Size_t_w_clone: calculate Parent size as observed parent size at time t plus sizes of all offspring observed at time t+1 (this ensures that a parent cannot have size equal to or less than offspring)
 	\item Clones_t: number of clones produced per parent
 	\item Loss_to_Offspring: number of segments lost to clonal offspring
	}
}
\description{
Determine parent plant size before segments were lost to offspring, the number of segments lost of offspring, the number of clonal offspring produced, and the first size observed for each clonal offspring.
}

