% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateNumberIndividualsperStage.R
\name{calculateNumberIndivperStage}
\alias{calculateNumberIndivperStage}
\title{Create Vector of the Number of Individuals per Stage}
\usage{
calculateNumberIndivperStage(TMdata, stages)
}
\arguments{
\item{TMdata}{Data used to create the transition matrix.}

\item{stages}{vector of stage classes.}
}
\description{
Create Vector of the Number of Individuals per Stage
}

