% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCactusDensityperPlot.R
\name{estimateCactusDensityperPlot}
\alias{estimateCactusDensityperPlot}
\title{Estimate cactus density per plot}
\usage{
estimateCactusDensityperPlot(Plant.Surveys.by.Year, Plot.Info)
}
\arguments{
\item{Plot_Info}{}
}
\description{
Estimate cactus density as well as *O. stricta* and *O. pusilla* density (number of segments) per plot.
}

