% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotElasticitySensitivity.R
\name{plotElasticitySensitivity}
\alias{plotElasticitySensitivity}
\title{Plot Elasticity and Sensitivity Data}
\usage{
plotElasticitySensitivity(dat, x_variable, y_variable, grouping_variable,
  facet_grid_var, xlab_label, ylab_label, n_transitions_per_stage,
  strip.text.x.size, strip.text.y.size, legend.text.size, axis.text.x.size,
  axis.title.x.size, axis.title.y.size, axis.text.y.size,
  strip.text.y.angle = 270,
  color_values = c(black.red(n_transitions_per_stage),
  black.yellow(n_transitions_per_stage),
  black.blue.orange(n_transitions_per_stage),
  black.red.yellow(n_transitions_per_stage),
  black.purple.yellow(n_transitions_per_stage),
  black.darkblue.yellow(n_transitions_per_stage),
  black.brown.yellow(n_transitions_per_stage),
  black.brown.grey(n_transitions_per_stage),
  black.blue.brown(n_transitions_per_stage)))
}
\arguments{
\item{dat}{Dataset}

\item{n_transitions_per_stage}{}
}
\description{
Plot Elasticity and Sensitivity Data
}

